feat: move SpanExt trait to bio_format module root
