+++ src/bio_format/mod.rs
@@ -1,3 +1,4 @@
+pub use nu_protocol::{Span, Value};
 /// SAM + BAM parsing facility.
 pub mod bam;
 /// BCF + VCF parsing facility.
@@ -17,4 +18,25 @@ pub mod gff;
 pub enum Compression {
     Uncompressed,
     Gzipped,
+}
+
+pub trait SpanExt {
+    fn with_string<S: ToString>(&self, s: S) -> Value;
+    fn with_string_or<S: ToString>(&self, s: Option<S>, default: &str) -> Value;
+}
+
+impl SpanExt for Span {
+    fn with_string<S: ToString>(&self, s: S) -> Value {
+        Value::String {
+            val: s.to_string(),
+            span: *self,
+        }
+    }
+
+    fn with_string_or<S: ToString>(&self, s: Option<S>, default: &str) -> Value {
+        Value::String {
+            val: s.map(|s| s.to_string()).unwrap_or(default.into()),
+            span: *self,
+        }
+    }
 }
